requirements.txt
appdirs==1.4.3
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.10.6
django-appconf==1.0.2
django-ckeditor==5.2.1
django-imagekit==4.0
django-import-export==0.5.1
gunicorn==19.7.1
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.0.0
pyparsing==2.2.0
six==1.10.0
tablib==0.11.4
whitenoise==3.3.0
